chore: put usecases in separate export file
