+++ lib/features/domain/usecases/usecases.dart
@@ -1,0 +1,5 @@
+export 'desk/create_desk_usecase.dart';
+export 'desk/get_all_desks_usecase.dart';
+export 'desk/get_desk_by_id_usecase.dart';
+export 'desk/update_desk_usecase.dart';
+export 'desk/delete_desk_usecase.dart';
