+++ test/features/presentation/bloc/desk/desk_bloc_test.dart
@@ -3,11 +3,7 @@ import 'package:deskify/core/core.dart';
 import 'package:deskify/features/data/models/desk_model.dart';
 import 'package:deskify/features/domain/entities/desk.dart';
 import 'package:deskify/features/domain/entities/preset.dart';
-import 'package:deskify/features/domain/usecases/create_desk_usecase.dart';
-import 'package:deskify/features/domain/usecases/delete_desk_usecase.dart';
-import 'package:deskify/features/domain/usecases/get_all_desks_usecase.dart';
-import 'package:deskify/features/domain/usecases/get_desk_by_id_usecase.dart';
-import 'package:deskify/features/domain/usecases/update_desk_usecase.dart';
+import 'package:deskify/features/domain/usecases/usecases.dart';
 import 'package:deskify/features/presentation/bloc/desk/desk_bloc.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
