+++ test/features/presentation/pages/home_page/home_page_test.dart
@@ -2,11 +2,7 @@ import 'package:dartz/dartz.dart';
 import 'package:deskify/features/domain/entities/desk.dart';
 import 'package:deskify/features/domain/entities/preset.dart';
 import 'package:deskify/features/domain/repository/desk_repository.dart';
-import 'package:deskify/features/domain/usecases/create_desk_usecase.dart';
-import 'package:deskify/features/domain/usecases/delete_desk_usecase.dart';
-import 'package:deskify/features/domain/usecases/get_all_desks_usecase.dart';
-import 'package:deskify/features/domain/usecases/get_desk_by_id_usecase.dart';
-import 'package:deskify/features/domain/usecases/update_desk_usecase.dart';
+import 'package:deskify/features/domain/usecases/usecases.dart';
 import 'package:deskify/features/presentation/bloc/desk/desk_bloc.dart';
 import 'package:deskify/features/presentation/pages/home_page/desk_carousel_slider.dart';
 import 'package:deskify/features/presentation/pages/home_page/desk_interaction_card.dart';
